Extract `FeedImageCell` helpers in their own file
