+++ EssentialFeed/EssentialFeediOSTests/Controllers/FeedViewControllerTests.swift
@@ -341,33 +341,3 @@ private extension FeedViewControllerTests {
         }
     }
 }
-
-private extension FeedImageCell {
-    func simulateRetryAction() {
-        feedImageRetryButton.simulateTap()
-    }
-
-    var isShowingLocation: Bool {
-        !locationContainer.isHidden
-    }
-
-    var isShowingImageLoadingIndicator: Bool {
-        feedImageContainer.isShimmering
-    }
-
-    var isShowingRetryAction: Bool {
-        !feedImageRetryButton.isHidden
-    }
-
-    var locationText: String? {
-        locationLabel.text
-    }
-
-    var descriptionText: String? {
-        descriptionLabel.text
-    }
-
-    var renderedImage: Data? {
-        feedImageView.image?.pngData()
-    }
-}
