+++ EssentialFeed/EssentialFeediOSTests/Controllers/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,39 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Stoyan Kostov on 21.06.23.
+//
+
+import UIKit
+import EssentialFeediOS
+
+extension FeedImageCell {
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+
+    var isShowingLocation: Bool {
+        !locationContainer.isHidden
+    }
+
+    var isShowingImageLoadingIndicator: Bool {
+        feedImageContainer.isShimmering
+    }
+
+    var isShowingRetryAction: Bool {
+        !feedImageRetryButton.isHidden
+    }
+
+    var locationText: String? {
+        locationLabel.text
+    }
+
+    var descriptionText: String? {
+        descriptionLabel.text
+    }
+
+    var renderedImage: Data? {
+        feedImageView.image?.pngData()
+    }
+}
